Date and time pickers work now.
They save locally, and time/date are displayed in list, though not perfect.

Text save does not work if the user has not selected a date or time, so to fix this, have an if null check or something?

Also need to have and option where the user doesn't need to choose a date or time, could do this by setting the date and time values to null first rather than to the current time?